requirements.txt
Flask
PyPDF2
python-docx
scikit-learn
sentence-transformers
torch
